Ensure that decls from referenced modules are fully scoped

+++ haskell/compiler/tests/test3/java-output/adl/test3/A.java
@@ -21,7 +21,7 @@ public class A {
   }
 
   public A() {
-    this.f_int = 0;
+    this.f_int = (short)0;
     this.f_string = "";
     this.f_bool = false;
   }
