+++ haskell/compiler/tests/test3/java-output/adl/test3/U.java
@@ -31,7 +31,7 @@ public class U {
 
   public U() {
     this.disc = Disc.F_INT;
-    this.value = 0;
+    this.value = (short)0;
   }
 
   public U(U other) {
