+++ haskell/compiler/tests/test5/java-output/adl/test5/U2.java
@@ -26,7 +26,7 @@ public class U2 {
 
   public U2() {
     this.disc = Disc.V;
-    this.value = 0;
+    this.value = (short)0;
   }
 
   public U2(U2 other) {
